1.0 Asset Validator
Migrating asset validator content from older repository. Compressed example content into an .unitypackage file and removed the uncompressed contents

Added license and license header file to project. Added MIT license to all files.

Modified the menu items to redirect to the editor window and modified their priority so that there is now a separator between several items. Modified the Asset Bundle and Resource Path validators and unit tests so that test contracts do not get swept up during normal usage. Removed menu item functionality for toggling off/on debugging.

Reorganized and cleaned up project now that this project will no longer have a premium and free component. Removed unused methods and classes

Updated readme with updated examples and the newer MIT license.

Altered progress shown to be a description of the time running rather than a percentage of the validation run, since it is a better way to show that the runner is actively working. Fixed a bug where the UI counts for info/warning/error logs are cut off when the count is three digits long.

Modified the editor window to disable all other controls when a validation run is being made. Added several helpful comments on AssetValidatorCI.Result members. Fixed a bug in the EnsureComponentIsUniqueValidator where it would flag a validation error anytime a targeted component is found rather than more than one. Modified the readme to show examples of continuous integration and screenshots of the editor tool.

Modified project to have an AssemblyDefinition file for runtime/editor code